+++ src/main.tsx
@@ -26,6 +26,7 @@ import SupportPage from './pages/SupportPage';
 import PrivacyPolicy from './components/privacy/privacyPolicy';
 import TOS from './pages/TOS';
 import Cookies from './pages/Cookies';
+import VerificationPage from './pages/VerificationPage';
 
 
 const router = createBrowserRouter([
@@ -66,10 +67,13 @@ const router = createBrowserRouter([
         element:  <Escrow/>
 
       },
-      
       {
         path: '/sign-up',
         element: <SignUpPage/>
+      },
+      {
+        path: '/verify-email',
+        element: <VerificationPage/>
       },
       {
         path: '/password-recovery',
